chore: move registration emails to its own file
